rename test_ray_cluster to test_ray_cluster_standalone
